+++ esperanto-tools-libs/src/RemoteProfiler.cpp
@@ -61,6 +61,8 @@ void RemoteProfiler::record(const ProfileEvent& event) {
     return;
   }
 
+  registerThread();
+
   auto cl = event.getClass();
   if ((cl != Class::ResponseReceived) && (cl != Class::CommandSent) && (cl != Class::DispatchEvent)) {
     // All other classes are also measured at the client, so here we avoid having both events in the client
@@ -80,6 +82,33 @@ void RemoteProfiler::record(const ProfileEvent& event) {
   }
 }
 
+void RemoteProfiler::sendProfilingEvent(Worker* worker, const ProfileEvent& event) {
+  assert(worker != nullptr);
+
+  auto threadId = event.getNumericThreadId();
+
+  // Check if the thread needs to be identified to the worker
+  SpinLock lock(workerAccessedThreadsMutex_);
+  auto& accessedThreads = workerAccessedThreads_[worker];
+  auto [it, created] = accessedThreads.emplace(threadId);
+  (void)it;
+  lock.unlock();
+
+  // Identify the thread to the worker if not already done
+  if (created) {
+    ProfileEvent identifyThreadEvent{Type::Instant, Class::IdentifyThread};
+    identifyThreadEvent.setThreadId(threadId);
+    auto threadName = getThreadName(threadId);
+    if (not threadName.empty()) {
+      identifyThreadEvent.setThreadName(std::move(threadName));
+    }
+    worker->sendProfilerEvent(identifyThreadEvent);
+  }
+
+  // Send the actual profiling event to the worker
+  worker->sendProfilerEvent(event);
+}
+
 void RemoteProfiler::recordRequestProfilingEvent(EventId eventId, StreamId streamId, const ProfileEvent& event) {
   auto worker = getWorker(streamId);
   if (worker == nullptr) {
@@ -88,7 +117,7 @@ void RemoteProfiler::recordRequestProfilingEvent(EventId eventId, StreamId strea
   }
 
   // Emit the request profiling event
-  worker->sendProfilerEvent(event);
+  sendProfilingEvent(worker, event);
 
   // Response profiling events do not have the stream id recorded
   SpinLock lock(eventToStreamAndEraliesMutex_);
@@ -96,7 +125,7 @@ void RemoteProfiler::recordRequestProfilingEvent(EventId eventId, StreamId strea
   if (it != earlyProfilingEvents_.end()) {
     // Try to process any response profiling event that might have arrived too early
     auto const& earlyEvent = it->second;
-    worker->sendProfilerEvent(earlyEvent);
+    sendProfilingEvent(worker, earlyEvent);
     earlyProfilingEvents_.erase(it);
   } else {
     // Save the event to stream association for the response profiling event
@@ -119,7 +148,7 @@ void RemoteProfiler::recordResponseProfilingEvent(EventId eventId, const Profile
     }
 
     lock.unlock();
-    worker->sendProfilerEvent(event);
+    sendProfilingEvent(worker, event);
   } else {
     // Delay this reponse profiling event since it arrived before its matching request profiling event
     auto [it2, inserted] = earlyProfilingEvents_.try_emplace(eventId, event);
@@ -129,6 +158,30 @@ void RemoteProfiler::recordResponseProfilingEvent(EventId eventId, const Profile
       RT_LOG(WARNING) << "Dropping remote profiling event with unknown worker since there is already another one "
                          "with the same event id";
     }
+  }
+}
+
+inline void RemoteProfiler::registerThread() {
+  auto threadId = std::this_thread::get_id();
+
+  SpinLock lock(threadIdNamesMutex_);
+  auto it = threadIdNames_.find(threadId);
+  if (it == threadIdNames_.end()) {
+    threadIdNames_.emplace(threadId, threadName_);
+  }
+}
+
+std::string RemoteProfiler::getThreadName(std::thread::id threadId) {
+  if (threadId == std::this_thread::get_id()) {
+    return threadName_;
+  }
+
+  SpinLock lock(threadIdNamesMutex_);
+  auto it = threadIdNames_.find(threadId);
+  if (it != threadIdNames_.end()) {
+    return it->second;
+  } else {
+    return std::string();
   }
 }